Feature: roomchat and unit test

+++ domains/auths/dto.go
@@ -1,6 +1,6 @@
 package auths
 
-import "goboilerplate/commons"
+import "chatroom-api/commons"
 
 type LoginRequest struct {
 	Email    string `json:"email" binding:"required"`
@@ -13,14 +13,14 @@ type RegisterRequest struct {
 }
 
 type LoginResponse struct {
-	ID     int                 `json:"uuid"`
+	ID     uint64              `json:"uuid"`
 	Status string              `json:"status"`
 	Token  commons.TokenDetail `json:"token"`
 	Email  string              `json:"email"`
 }
 
 type RegisterResponse struct {
-	ID     int                 `json:"uuid"`
+	ID     uint64              `json:"uuid"`
 	Email  string              `json:"email"`
 	Status string              `json:"status"`
 	Token  commons.TokenDetail `json:"token"`
